+++ src/Messaging/Interceptors/CachingInterceptor.cs
@@ -23,41 +23,14 @@ public sealed class CachingInterceptor : DefaultCachingInterceptor
         ILogger<CachingInterceptor> logger,
         ITypeEncoder typeEncoder,
         IStorage<IAbstractMessage, CachingResult> cache,
-        INamedOptions<MessagingClientOptions> options) : base(logger, typeEncoder, cache, options) { }
+        INamedOptions<MessagingClientOptions> options) : base((ILogger)logger, typeEncoder, cache, options) { }
 
     /// <inheritdoc/>
-    public override async Task<object> Intercept(MessageInterceptor next, IAbstractMessage message, CancellationToken token)
+    protected override async ValueTask<object> InterceptInternal(SharedMessageHandler next, IAbstractMessage message, CancellationToken token)
     {
         if(message is INonCaching)
             return await next(message, token);
 
-        return await base.Intercept(next, message, token);
+        return await base.InterceptInternal(next, message, token);
     }
-}
-
-/// <summary>
-///     Message response (including failures) caching interceptor.
-/// </summary>
-/// <remarks>
-///     The interceptor depends on <see cref="MessagingClientOptions.TransientExceptions"/>,
-///     <see cref="INonCaching"/>.
-/// </remarks>
-public sealed class CachingInterceptor<TMessage, TResponse> : IMessageInterceptor<TMessage, TResponse>
-    where TMessage : IMessage<TResponse>
-{
-    private readonly CachingInterceptor interceptor;
-
-    /// <summary/>
-    public CachingInterceptor(
-        ILogger<CachingInterceptor> logger,
-        ITypeEncoder typeEncoder,
-        IStorage<IAbstractMessage, CachingResult> cache,
-        INamedOptions<MessagingClientOptions> options)
-    {
-        interceptor = new CachingInterceptor(logger, typeEncoder, cache, options);
-    }
-
-    /// <inheritdoc/>
-    public async Task<TResponse> Intercept(MessageInterceptor<TMessage, TResponse> next, TMessage message, CancellationToken token) =>
-        (TResponse)await interceptor.Intercept(async (m, t) => (await next((TMessage)m, t))!, message, token);
 }