+++ tests/Messaging.Tests/Interceptors/CancellationDelayInterceptorTests.cs
@@ -1,5 +1,4 @@
 ﻿using Assistant.Net.Abstractions;
-using Assistant.Net.Messaging.Abstractions;
 using Assistant.Net.Messaging.Interceptors;
 using Assistant.Net.Messaging.Options;
 using Assistant.Net.Messaging.Tests.Mocks;
@@ -16,7 +15,7 @@ namespace Assistant.Net.Messaging.Tests.Interceptors;
 public class CancellationDelayInterceptorTests
 {
     [Test]
-    public async Task Intercept_delaysCancellation_longRunningOperation()
+    public async Task Intercept_delaysCancellation_requestAndLongRunningOperation()
     {
         var cancelledSource = new CancellationTokenSource(0);
         var timer = Stopwatch.StartNew();
@@ -32,35 +31,62 @@ public class CancellationDelayInterceptorTests
     }
 
     [Test]
-    public async Task Intercept_doesNotDelay_noCancellation()
+    public async Task Intercept_doesNotDelay_requestAndNoCancellation()
     {
         var source = new CancellationTokenSource();
         var timer = Stopwatch.StartNew();
 
-        await Interceptor.Intercept((_, token) => Task.Run<object>(() => Response, token), Message, source.Token);
+        await Interceptor.Intercept((_, _) => ValueTask.FromResult<object>(Response), Message, source.Token);
         timer.Stop();
 
         timer.Elapsed.Should().BeLessThan(Options.CancellationDelay);
     }
 
-    [SetUp]
+    [Test]
+    public async Task Intercept_delaysCancellation_publishAndLongRunningOperation()
+    {
+        var cancelledSource = new CancellationTokenSource(0);
+        var timer = Stopwatch.StartNew();
+
+        await Interceptor.Awaiting(x => x.Intercept(async (_, token) =>
+        {
+            await Task.WhenAll(Task.Delay(Timeout.Infinite, token));
+        }, Message, cancelledSource.Token)).Should().ThrowAsync<OperationCanceledException>();
+        timer.Stop();
+
+        timer.Elapsed.Should().BeGreaterThan(Options.CancellationDelay);
+    }
+
+    [Test]
+    public async Task Intercept_doesNotDelay_publishAndNoCancellation()
+    {
+        var source = new CancellationTokenSource();
+        var timer = Stopwatch.StartNew();
+
+        await Interceptor.Intercept((_, _) => ValueTask.CompletedTask, Message, source.Token);
+        timer.Stop();
+
+        timer.Elapsed.Should().BeLessThan(Options.CancellationDelay);
+    }
+
+    [OneTimeSetUp]
     public void Setup()
     {
         Options = new MessagingClientOptions {CancellationDelay = TimeSpan.FromSeconds(0.1)};
         var services = new ServiceCollection()
             .AddLogging()
             .AddTypeEncoder()
-            .AddSingleton<INamedOptions<MessagingClientOptions>>(new TestNamedOptions {Value = Options})
+            .AddSingleton<INamedOptions<MessagingClientOptions>>(new TestNamedOptions(Options))
             .AddSingleton<CancellationDelayInterceptor>()
             .AddTransient<CachingInterceptor>();
         Provider = services.BuildServiceProvider();
     }
 
-    [TearDown]
+    [OneTimeTearDown]
     public void TearDown() => Provider.Dispose();
 
     private ServiceProvider Provider { get; set; } = default!;
-    private IAbstractInterceptor Interceptor => Provider.GetRequiredService<CancellationDelayInterceptor>();
+    private CancellationDelayInterceptor Interceptor => Provider.GetRequiredService<CancellationDelayInterceptor>();
     private MessagingClientOptions Options { get; set; } = default!;
     private static TestMessage Message => new(0);
     private static TestResponse Response => new(false);